# Add deepequal check for objectWatcher.Update()