@RainbowMango 2025-03-10
> If https://github.com/karmada-io/karmada/pull/6148 can be merged and this PR can be rebased on it, the changes brought by this PR can be more clearly shown through these metrics.

FYI. The #6148 has been merged.