@zach593 2025-03-10
> FYI. The #6148 has been merged.

Got it, I just rebased the commit. It may need in-depth testing, I'll post the test results ASAP.