@zach593 2025-03-11
test comparison

setup:

- run `hack/local-up-karmada.sh` in current master branch
- apply yaml files in `samples/nginx/`
- restart karmada-controller-manager

without this commit:

![image](https://github.com/user-attachments/assets/0b44ef37-5743-4934-b9ec-92e379470b28)
![image](https://github.com/user-attachments/assets/b161fbc3-ed17-436d-92b4-348da6b50ad4)

with this commit:

![image](https://github.com/user-attachments/assets/48af6cce-69a1-4482-9bbb-e3c593d70e30)
![image](https://github.com/user-attachments/assets/2de3e297-3c45-4157-bad8-14f0c0bdf5bd)