@zach593 2025-03-19
after updated to [a472610](https://github.com/karmada-io/karmada/commit/a4726100e874b7e30c5167608d28c073988e3726) , the equality check still could work.

![image](https://github.com/user-attachments/assets/ba2fd02e-529e-484c-af73-ada9165dda4c)
![image](https://github.com/user-attachments/assets/4abfbdd4-b5e3-4152-8c9d-e7052e5f9b60)