@zach593 2025-04-15
> Do we have a test report? I believe this patch would improve performance significantly, it should be high-lighted in release notes.

I'll make one this week.

This PR can theoretically reduce the performance pressure of the member cluster apiserver and reduce the average time cost of the `execution controller` and `work status controller`.
Based on #6192, if the objectWatcher is limited by the rate limiter, it can also reduce this limitation. However, because #6192 has not been merged yet, that part is not suitable to be regarded as an optimization brought by this PR. :smile: