@zach593 2025-04-26
### Performance Test Report

To reduce the influence of other controllers on work-related controllers and to extend the processing time for observation purposes, a rather extreme configuration was used:

![image](https://github.com/user-attachments/assets/6b1c0108-1089-4d9d-82cc-1d83431e76b1)

The test data consisted of 20,000 Deployments distributed across 2 member clusters.

The impact of this commit:

- During startup, no HTTP Update requests will be sent to member clusters anymore.
- The CPU usage of the member clusters’ kube-apiserver can be reduced from 2 cores to 1.5 cores.
- The reconciliation time cost of execution controller can be reduced from 3 ms to 556 µs, a decrease of about 80%.
- It ensures that the work-status controller works properly during startup, instead of failing and placing items into the rate limit queue (seems we accidentally fix a bug).
- The CPU usage of the karmada controller-manager has increased slightly, which is reasonable due to the introduction of `DeepEqual`.
- The CPU usage of the karmada-apiserver has also increased. It is suspected to be related to the fact that the work-status controller is now functioning properly. Further verification is needed.

<details>
<summary> Click to view detailed report </summary>

### Without this commit
It took 2 minutes to clear the queue.
For the reconciliation time per execution controller, the P90 and average were: 9.19ms, 2.98ms.

![image](https://github.com/user-attachments/assets/3a99c0f5-a153-431c-b9e8-135e8849efd7)

However, it can be observed that the work-status controller cleared its queue within 30 seconds, but there were still a large number of items in the rate limit queue.
The reason is that the work-status processing failed at the beginning when the controller just started.

![image](https://github.com/user-attachments/assets/a634e563-a660-4065-a254-e6ecae2af4c6)

In `objectWatcher.Update()`, the update operations on the member clusters did not cause any actual modifications.

![image](https://github.com/user-attachments/assets/2ddbd885-0307-4dc0-b95d-4f565aa29ceb)

During startup:
- Member clusters’ kube-apiserver CPU peaked at 0.5 cores, memory usage peaked at 2GB.
- In the karmada cluster, the controller-manager's CPU peaked at 0.9 cores, memory usage peaked at 4GB.
- The karmada-apiserver's CPU peaked at 0.5 cores, and memory usage peaked at 9.5GB.

![image](https://github.com/user-attachments/assets/b003d3cc-1dd4-4541-88fb-10a33bf6ce19)

### With this commit

It took 3 minutes to clear the queue.  
For the reconciliation time per execution controller, the P90 and average were: 915 µs, 556 µs.

It is obvious that the work-status controller was functioning properly this time, leading to an increase in processing time. However, it appears that the execution controller's queue was cleared within just 20 seconds.

![image](https://github.com/user-attachments/assets/bacb413a-923c-49de-a496-6b9f5c174905)

In `objectWatcher.Update()`, no HTTP update requests to the member clusters were triggered.
![image](https://github.com/user-attachments/assets/de426163-a7d7-4576-85f7-a2503d2e5b03)

During startup:
- Member clusters’ kube-apiserver CPU peaked at 0.15 cores, memory usage peaked at 2GB.
- In the karmada cluster, the controller-manager's CPU peaked at 1.1 cores, memory usage peaked at 4.2GB.
- The karmada-apiserver's CPU peaked at 1.3 cores, and memory usage peaked at 9.2GB.

![image](https://github.com/user-attachments/assets/b3f1c1d9-ef62-4f77-82e9-2a844d4942c2)

</details>